+++ plugins/ps-demo4/CHANGELOG.md
@@ -1,4 +1,12 @@
 
+## 0.1.0
+### Added
+- Add `peripheral-updated` event to DATA packet in tcp-daemon
+
+### Changed
+- Use metadata of DATA packet with `peripheral-updated` event to update peripheral object
+- Keep pid for senor updates
+
 ## 0.0.3
 ### Changed
 - improve documentation and verbose messages in the shortcut script for running SensorWeb